pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "spot-train"
version = "0.1.0"
description = "Spec-driven MVP for memory-backed Spot task orchestration."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
  { name = "spot-train contributors" }
]
dependencies = [
  "pydantic>=2.7,<3",
  "PyYAML>=6,<7",
]

[project.optional-dependencies]
dev = [
  "pytest>=8,<9",
  "ruff>=0.4.0,<1",
]
repl = [
  "cmd2==3.4.0",
]
strands = [
  "strands-agents==1.34.1",
  "strands-agents-tools==0.3.0",
]
bedrock = [
  "boto3>=1.26,<2",
  "botocore>=1.29,<2",
]
spot = [
  "bosdyn-client==5.1.4",
  "bosdyn-mission==5.1.4",
]
perception = [
  "opencv-python>=4.8,<5",
  "numpy>=1.24,<3",
]

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
  "E",
  "F",
  "I",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"
